feat(athenaeum): bigger stars + half-star ratings across all surfaces
Stars were 1rem Unicode buttons rendering integers only. Bumped to
1.4rem (1.6rem on book-edit, 1.2rem on /recommendations) and now
support 0.5-step ratings everywhere they appear.

Implementation:
- New star-helpers.ts: bookStarsHtml / seriesStarsHtml emit a single
  ★ button per slot; ratingFromClick picks 0.5 or 1 from the click X
  relative to the star's bounding box; previewHover updates
  data-hover-fill for live half/full preview during cursor hover.
- CSS paints each star via linear-gradient + background-clip:text on
  the glyph itself — no overlay DOM. --fill-pct is driven by
  data-fill / data-hover-fill (empty / "half" / "full"). Same trick on
  .card-star, .series-card-star, .rec-star, .rating-stars__btn.
- Hover preview now half-fills the hovered star based on cursor X.
- Wired in book-card + book-grid (library/favs/search), series-list
  (both /series page and /library group headings), recommendations
  card, and the book-edit rating control (which gained click-to-clear).
- Backend: BookUpdate.rating + RecommendationAddPayload.rating + the
  add_book helper signature now accept floats. Series rating was
  already float. DB column is REAL (no migration needed).

Bumps style.css?v=34 and app.js?v=63.

+++ books/helpers/db.py
@@ -1939,7 +1939,7 @@ def insert_book(
     tags: list[str] | None,
     date_added: str,
     date_finished: str | None,
-    rating: int | None,
+    rating: float | None,
     reading_status: str = "unread",
     progress: float | None = None,
     is_favorite: int = 0,
